Clear and change directory configuration for restarting the training process now present...
Clear and change directory configuration for restarting the training process now present. The model is now saved to the current directory in folder modelbase, for deletion for next model training session. The os call is positioned above the transformer import for a reason. It won't work if it's under that import... No idea why that's the case though. It just appears to be and people mention it on the web. Now this feature works perfectly after discovering that too!