Updated with an undo feature bound to the 'u' key.
Updated with an undo feature bound to the 'u' key for the Stockfish version. I'm writing a Shredder version of this script too which isn't finished, but I've included it in the submission. The autoplay and movement feature is finished, but the drag and drop system is stubborn for some reason using this engine. Hope to update the shredder version soon though to be fully functional like the Stockfish version. Shredder clearly has better ai than Stockfish too. It's a much more incredible best chess system.